Add output to solution class, prepare single step
Add a member function to the solution class that adds output capability
compatible with the VisClaw package from Clawpack.  Added some
functionality to the Grid and Solver classes to help with evolving
a solution a single time step.  Basic test is written and the Solver
member function step needs to be implemented to complete the test.